test/reftest/web-platform-tests/dom_converted/events/EventTarget-dispatchEvent-returnvalue.html

<!DOCTYPE html>
<meta charset=utf-8>
<!--EDIT-->
<!--title>EventTarget.dispatchEvent: return value</title>
<link rel="help" href="https://dom.spec.whatwg.org/#concept-event-dispatch">
<link rel="help" href="https://dom.spec.whatwg.org/#dom-event-preventdefault">
<link rel="help" href="https://dom.spec.whatwg.org/#dom-event-defaultprevented"-->
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  #table {
    display: none;
  }
  .col {
    display: inline-block;
  }
  .cell {
    border: 1px solid black;
    padding: 0;
    margin: 1px;
  }
</style>
<div id=log></div>
<div id="table">
  <div id="table-body">
    <div class="col" id="parent">
      <p class="cell">Shady Grove</p>
      <p class="cell" id="target">Over the river, Charlie</p>
    </div>
    <div class="col">
      <p class="cell">Aeolian</p>
      <p class="cell">Dorian</p>
    </div>
  </div>
</div>
<script>
test(function() {
    var event_type = "foo";
    var target = document.getElementById("target");
    var parent = document.getElementById("parent");
    var default_prevented;

    parent.addEventListener(event_type, function(e) {}, true);
    target.addEventListener(event_type, function(e) {
        evt.preventDefault();
        default_prevented = evt.defaultPrevented;
    }, true);
    target.addEventListener(event_type, function(e) {}, true);

    // EDIT
    // var evt = document.createEvent("Event");
    // evt.initEvent(event_type, true, true);
    var evt = new Event(event_type, {"bubbles":true, "cancelable":true});

    assert_true(parent.dispatchEvent(evt));
    assert_false(target.dispatchEvent(evt));
    assert_true(default_prevented);
}, "Return value of EventTarget.dispatchEvent.");
</script>
